chunk: assume generated if TerrainGenerated tag is not present
vanilla (and modern PM) don't have a TerrainGenerated tag - it's unique to legacy PM. Therefore, the presence of a chunk which does not have a TerrainGenerated tag implies that the chunk is generated already, because vanilla and PM3 don't save non-generated chunks in the first place.

This code was causing problems with imported worlds, because while
non-generated chunks are not saved, generated (but not populated) chunks
ARE saved by both PM3 and vanilla. Therefore, entering into such a chunk
would lead to the server incorrectly assuming the chunk wasn't
generated, and preventing the player from entering it.